tests/wpt/mozilla/tests/mozilla/document-compatMode-loose.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script>
            test(function() {
                assert_equals(document.compatMode, "BackCompat", "test1-0, Loose HTML");
            });
        </script>
    </head>
</html>
